OSIS-1237 Fixed error when submitting partim form
	modified:   base/forms/learning_unit/learning_unit_create.py
	modified:   base/forms/learning_unit/learning_unit_create_2.py
