+++ base/forms/learning_unit/learning_unit_create.py
@@ -30,8 +30,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from base.forms.utils.acronym_field import AcronymField, PartimAcronymField, split_acronym
 from base.forms.utils.choice_field import add_blank
-from base.models import entity_version, academic_year
-from base.models.academic_year import MAX_ACADEMIC_YEAR_FACULTY, MAX_ACADEMIC_YEAR_CENTRAL
+from base.models import entity_version
 from base.models.campus import find_main_campuses
 from base.models.entity_component_year import EntityComponentYear
 from base.models.entity_container_year import EntityContainerYear
@@ -127,8 +126,6 @@ class LearningUnitYearModelForm(forms.ModelForm):
 
         if kwargs.get('instance'):
             self.fields['academic_year'].disabled = True
-        else:
-            self._restrict_academic_years_choice()
 
     class Meta:
         model = LearningUnitYear
@@ -144,15 +141,6 @@ class LearningUnitYearModelForm(forms.ModelForm):
                     max_value=MAXIMUM_CREDITS)
             }
         }
-
-    def _restrict_academic_years_choice(self):
-        current_academic_year = academic_year.current_academic_year()
-        end_year_range = MAX_ACADEMIC_YEAR_FACULTY if self.person.is_faculty_manager() else MAX_ACADEMIC_YEAR_CENTRAL
-
-        self.fields["academic_year"].queryset = academic_year.find_academic_years(
-            start_year=current_academic_year.year,
-            end_year=current_academic_year.year + end_year_range
-        )
 
     # TODO :: Move assignment to self.instance from save into __init__
     # TODO :: Make these kwarg to args (learning_container_year, learning_unit, ... are required args)
